successful implementation of v1 and v2 deploy; v2 release
